# src/web/api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Container Farm Control API
  version: "1.0"
servers:
  - url: http://localhost:5000
paths:
  /api/status:
    get:
      summary: Get basic system status
      responses:
        '200':
          description: Status response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
      security:
        - ApiKeyAuth: []
  /api/control/{output}:
    post:
      summary: Control a specific output
      parameters:
        - in: path
          name: output
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  default: toggle
      responses:
        '200':
          description: Control response
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                  action:
                    type: string
                  success:
                    type: boolean
      security:
        - ApiKeyAuth: []
  /api/students:
    get:
      summary: List registered students
      responses:
        '200':
          description: Student list
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      type: object
    post:
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                grade:
                  type: string
                id:
                  type: string
      responses:
        '200':
          description: Added student
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  student:
                    type: object
      security:
        - ApiKeyAuth: []
  /api/activities:
    post:
      summary: Log a student activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                student:
                  type: string
                action:
                  type: string
                details:
                  type: string
      responses:
        '200':
          description: Activity logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  entry:
                    type: object
      security:
        - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key